Bundle Viewing Functions
Created page components to direct organizers and collect sheet URLs.

Added functions to handle user input and submission.

Created variables to hold sheet data and information about the data.

Created functions to parse fetched data and variables to hold essential information.

Added functions to properly display an image that was uploaded to a google form.

Added functions for quick expansion and collapsing of all summary and detail elements on a page.